requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI & ML
google-generativeai==0.3.0
openai==1.3.0  # Optional for fallback
langchain==0.0.340
langchain-community==0.0.10

# Data Processing
pandas==2.1.3
numpy==1.24.3
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.1

# Utilities
pydantic[email]==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2
redis==5.0.1
celery==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# API Clients
google-api-python-client==2.108.0
newsapi-python==0.2.7
yelpapi==3.0.0